requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
redis>=4.0.0
torch==1.9.0
transformers==4.11.3
numpy==1.21.2
prometheus-client==0.11.0
language-tool-python==2.7.1
nltk==3.6.3
textstat>=0.7.3
pydantic==1.8.2
python-multipart>=0.0.5
aiohttp>=3.8.0
asyncio>=3.4.3
peft==0.8.2
bitsandbytes==0.42.0
sentencepiece==0.1.99
protobuf==4.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sentence-transformers==2.1.0
scikit-learn==0.24.2 